Preserve dtype in simple modis interpolation to avoid unnecessary casting

+++ geotiepoints/simple_modis_interpolator.py
@@ -151,7 +151,7 @@ def interpolate_geolocation_cartesian(lon_array, lat_array, res_factor=4):
     x, y = np.meshgrid(x, y)
     coordinates = np.array([y, x])  # Used by map_coordinates, major optimization
 
-    new_x = np.empty((num_rows * res_factor, num_cols * res_factor), dtype=np.float64)
+    new_x = np.empty((num_rows * res_factor, num_cols * res_factor), dtype=lon_array.dtype)
     new_y = new_x.copy()
     new_z = new_x.copy()
     nav_arrays = [(x_in, new_x), (y_in, new_y), (z_in, new_z)]
